+++ backend/uniqon/src/main/java/com/ssafy/uniqon/service/startup/qna/StartupQuestionService.java
@@ -13,29 +13,29 @@ import com.ssafy.uniqon.exception.ex.ErrorCode;
 import com.ssafy.uniqon.repository.startup.qna.StartupQuestionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.ssafy.uniqon.exception.ex.ErrorCode.INVALID_ACCESS_MEMBER;
 import static com.ssafy.uniqon.exception.ex.ErrorCode.QUESTION_NOT_FOUND;
 
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 @Service
 public class StartupQuestionService {
-
     private final StartupQuestionRepository startupQuestionRepository;
-    private final StartupAnswerService startupAnswerService;
 
     @Transactional
-    public Long 질문등록(Long memberId, StartupQuestionReqDto startupQuestionReqDto) {
+    public Long 질문등록(Long memberId, Long startupId, StartupQuestionReqDto startupQuestionReqDto) {
         Member member = new Member();
         member.changeId(memberId);
 
         Startup startup = new Startup();
-        startup.changeId(startupQuestionReqDto.getStartupId());
+        startup.changeId(startupId);
 
         StartupQuestion startupQuestion = StartupQuestion.builder()
                 .member(member)
@@ -48,17 +48,7 @@ public class StartupQuestionService {
     }
 
     public List<StartupQuestionResDto> 질문조회(Long memberId, Long startupId) {
-//        List<StartupQuestion> startupQuestionList = startupQuestionRepository.findAllByStartupId(startupId);
-//        List<StartupQuestionResDto> startupQuestionResDtoList = startupQuestionList.stream()
-//                .map(startupQuestion -> new StartupQuestionResDto(startupQuestion))
-//                .collect(Collectors.toList());
-//
-//        startupQuestionResDtoList.forEach(startupQuestionResDto -> {
-//            List<AnswerParentResponseDto> answerParentResponseDtoList = startupAnswerService.댓글조회(startupQuestionResDto.getStartupQuestionId());
-//            startupQuestionResDto.changeParentResponseDto(answerParentResponseDtoList);
-//        });
         List<StartupQuestionResDto> startupQuestionResDtoList = startupQuestionRepository.findStartupQuestionResDtoList(memberId, startupId);
-
         return startupQuestionResDtoList;
     }
 
@@ -68,13 +58,28 @@ public class StartupQuestionService {
         Long lastIdOfList = questions.isEmpty() ?
                 null : questions.get(questions.size() - 1).getStartupQuestionId();
 
-        return new CursorResult(questions, hasNext(startupId, lastIdOfList), lastIdOfList);
+         return new CursorResult(questions, hasNext(startupId, lastIdOfList), lastIdOfList);
     }
 
     private List<StartupQuestionResDto> getQuestions(Long memberId, Long startupId, Long cursorId, Pageable page) {
         return cursorId == null ?
                 this.startupQuestionRepository.findQuestionListDtoPage(memberId, startupId, page) :
                 this.startupQuestionRepository.findQuestionListDtoLessPage(memberId, startupId, cursorId, page);
+    }
+
+    public CursorResult<StartupQuestionResDto> 질문조회페이징_PUBLIC(Long startupId, Long cursorId, Pageable page) {
+
+        List<StartupQuestionResDto> questions = getQuestions_PUBLIC(startupId, cursorId, page);
+        Long lastIdOfList = questions.isEmpty() ?
+                null : questions.get(questions.size() - 1).getStartupQuestionId();
+
+        return new CursorResult(questions, hasNext(startupId, lastIdOfList), lastIdOfList);
+    }
+
+    private List<StartupQuestionResDto> getQuestions_PUBLIC(Long startupId, Long cursorId, Pageable page) {
+        return cursorId == null ?
+                this.startupQuestionRepository.findQuestionListDtoPage(startupId, page) :
+                this.startupQuestionRepository.findQuestionListDtoLessPage(startupId, cursorId, page);
     }
 
     private Boolean hasNext(Long startupId, Long cursorId) {
@@ -92,14 +97,24 @@ public class StartupQuestionService {
     }
 
     @Transactional
-    public void 질문수정(StartupQuestionUpdateReqDto startupQuestionUpdateReqDto) {
-        StartupQuestion startupQuestion = startupQuestionRepository.findById(startupQuestionUpdateReqDto.getStartupQuestionId())
+    public void 질문수정(Long memberId, Long startupQuestionId, StartupQuestionUpdateReqDto startupQuestionUpdateReqDto) {
+        StartupQuestion startupQuestion = startupQuestionRepository.findById(startupQuestionId)
                 .orElseThrow(() -> new CustomException(QUESTION_NOT_FOUND));
-        startupQuestion.update(startupQuestionUpdateReqDto);
+        if (startupQuestion.getMember().getId().equals(memberId)) {
+            startupQuestion.update(startupQuestionUpdateReqDto);
+        } else {
+            throw new CustomException(INVALID_ACCESS_MEMBER);
+        }
     }
 
     @Transactional
-    public void 질문삭제(Long startupQuestionId) {
-        startupQuestionRepository.deleteById(startupQuestionId);
+    public void 질문삭제(Long memberId, Long startupQuestionId) {
+        StartupQuestion startupQuestion = startupQuestionRepository.findById(startupQuestionId)
+                .orElseThrow(() -> new CustomException(QUESTION_NOT_FOUND));
+        if (startupQuestion.getMember().getId().equals(memberId)) {
+            startupQuestionRepository.deleteById(startupQuestionId);
+        } else {
+            throw new CustomException(INVALID_ACCESS_MEMBER);
+        }
     }
 }
